Prepare drone

+++ devboost.dronedelivery.felipe/devboost.dronedelivery.felipe.EF/Entities/PedidoDrone.cs
@@ -9,7 +9,7 @@ namespace devboost.dronedelivery.felipe.EF.Entities
         public Drone Drone { get; set; }
         public int PedidoId { get; set; }
         public Pedido Pedido { get; set; }
-        public int Distancia { get; set; }
+        public double Distancia { get; set; }
         public int StatusEnvio { get; set; }
         public DateTime DataHoraFinalizacao { get; set; }
     }
